requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.10
PyYAML>=6.0
joblib>=1.3
matplotlib>=3.7
seaborn>=0.13
tqdm>=4.66
lightgbm>=4.0
transformers>=4.40
torch>=2.1
modelscope>=1.18
